# docs/api-contract/api-contract.yml
openapi: 3.0.3
info:
  title: EdTech AI Platform API
  version: 1.1.0
  description: |
    Historical OpenAPI draft for Backend + AI Service.
    Current public client contract is generated by API Gateway Swagger at GET /api/docs
    and summarized in docs/api-contract/gateway.md.
    All endpoint descriptions map to User Stories (US-001..US-052).
    Standard response envelope for all APIs:
    {
      "success": true|false,
      "data": object|null,
      "meta": { "requestId": "uuid", "timestamp": "ISO-8601", "pagination": {...optional} },
      "error": { "code": "...", "message": "...", "details": [...] } | null
    }
servers:
  - url: https://api.edtech.local/api
    description: API Gateway public ingress
  - url: https://ai.edtech.local/v1
    description: AI Service internal/specialized service surface
...
security:
  - bearerAuth: []
paths:
...
  /classes:
    post:
      tags: [Course/Class APIs]
      summary: Create class
      description: "US-004"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateClassRequest' }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/ClassroomResponse' } } } }
    get:
      tags: [Course/Class APIs]
      summary: List classes
      description: "US-004, US-042"
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/PaginatedClassesResponse' } } } }
  /classes/{classId}/invites:
    post:
      tags: [Course/Class APIs]
      summary: Invite students to class
      description: "US-004"
      parameters:
        - $ref: '#/components/parameters/ClassId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [emails]
              properties:
                emails:
                  type: array
                  items: { type: string, format: email }
      responses:
        '202':
          description: Async invite dispatch
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JobAcceptedResponse' }
  /courses:
    post:
      tags: [Course/Class APIs]
      summary: Create course
      description: "US-022, US-024"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Course' }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/CourseResponse' } } } }
    get:
      tags: [Course/Class APIs]
      summary: List courses
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/PaginatedCoursesResponse' } } } }
...
  /chat/sessions:
    post:
      tags: [Chat / AI APIs]
      summary: Create chat session
      description: "US-006, US-009"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                language: { type: string, default: vi }
      responses:
        '201':
          description: Created
          content: { application/json: { schema: { $ref: '#/components/schemas/ChatSessionResponse' } } }
    get:
      tags: [Chat / AI APIs]
      summary: List chat sessions
      description: "US-014"
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/PaginatedChatSessionsResponse' } } }
  /chat/sessions/{sessionId}/messages:
    post:
      tags: [Chat / AI APIs]
      summary: Send message to AI orchestrator
      description: "US-006, US-007, US-008, US-011, US-012, US-013, US-015, US-016, US-023"
      parameters:
        - $ref: '#/components/parameters/SessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChatMessageCreate' }
      responses:
        '200':
          description: AI response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChatResponseEnvelope' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
    get:
      tags: [Chat / AI APIs]
      summary: Get chat history
      description: "US-014"
      parameters:
        - $ref: '#/components/parameters/SessionId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaginatedChatMessagesResponse' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page: { in: query, name: page, schema: { type: integer, minimum: 1, default: 1 } }
    Limit: { in: query, name: limit, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
...
    ClassId: { in: path, name: classId, required: true, schema: { type: string, format: uuid } }
    SessionId: { in: path, name: sessionId, required: true, schema: { type: string, format: uuid } }
...
  responses:
...
    UnprocessableEntity:
      description: Unprocessable content
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiErrorResponse' } } }
  schemas:
    ResponseMeta:
      type: object
      required: [requestId, timestamp]
      properties:
        requestId: { type: string, format: uuid }
        timestamp: { type: string, format: date-time }
        pagination:
          type: object
          properties:
            page: { type: integer }
            limit: { type: integer }
            total: { type: integer }
    ApiError:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        details: { type: array, items: { type: string } }
    ApiErrorResponse:
      type: object
      required: [success, data, meta, error]
      properties:
        success: { type: boolean, example: false }
        data: { type: object, nullable: true, example: null }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
        error: { $ref: '#/components/schemas/ApiError' }
...
    Course:
      type: object
      required: [id, name]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
        teacherId: { type: string, format: uuid }
    Classroom:
      type: object
      required: [id, name, teacherId]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        courseId: { type: string, format: uuid }
        teacherId: { type: string, format: uuid }
        inviteCode: { type: string }
    CreateClassRequest:
      type: object
      required: [name, courseId]
      properties:
        name: { type: string, minLength: 2 }
        courseId: { type: string, format: uuid }
...
    ChatSession:
      type: object
      properties:
        id: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        language: { type: string }
    ChatMessage:
      type: object
      properties:
        id: { type: string, format: uuid }
        sessionId: { type: string, format: uuid }
        role: { type: string, enum: [user, assistant, system] }
        content: { type: string }
        createdAt: { type: string, format: date-time }
    ChatMessageCreate:
      type: object
      required: [content]
      properties:
        content: { type: string, minLength: 1, maxLength: 8000 }
        language: { type: string }
        imageUrl: { type: string, nullable: true }
    ChatResponse:
      type: object
      properties:
        message: { $ref: '#/components/schemas/ChatMessage' }
        intent: { type: string }
        confidence: { type: number }
        citations:
          type: array
          items:
            type: object
            properties:
              source: { type: string }
              snippet: { type: string }
        followUpQuestion: { type: string, nullable: true }
...
    JobAcceptedResponse:
      type: object
      required: [jobId, status]
      properties:
        jobId: { type: string, format: uuid }
        status: { type: string, enum: [queued] }
...
    PaginatedClasses:
      type: object
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/Classroom' } }
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
    PaginatedCourses:
      type: object
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/Course' } }
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
...
    PaginatedChatSessions:
      type: object
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/ChatSession' } }
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
    PaginatedChatMessages:
      type: object
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/ChatMessage' } }
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
...
    CourseResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        data: { $ref: '#/components/schemas/Course' }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
        error: { type: object, nullable: true, example: null }
    ClassroomResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        data: { $ref: '#/components/schemas/Classroom' }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
        error: { type: object, nullable: true, example: null }
...
    ChatSessionResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        data: { $ref: '#/components/schemas/ChatSession' }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
        error: { type: object, nullable: true, example: null }
    ChatResponseEnvelope:
      type: object
      properties:
        success: { type: boolean, example: true }
        data: { $ref: '#/components/schemas/ChatResponse' }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
        error: { type: object, nullable: true, example: null }
...
    PaginatedClassesResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        data: { type: array, items: { $ref: '#/components/schemas/Classroom' } }
        meta:
          allOf:
            - $ref: '#/components/schemas/ResponseMeta'
            - type: object
              properties:
                pagination: { $ref: '#/components/schemas/PaginatedClasses' }
        error: { type: object, nullable: true, example: null }
    PaginatedCoursesResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        data: { type: array, items: { $ref: '#/components/schemas/Course' } }
        meta:
          allOf:
            - $ref: '#/components/schemas/ResponseMeta'
            - type: object
              properties:
                pagination: { $ref: '#/components/schemas/PaginatedCourses' }
        error: { type: object, nullable: true, example: null }
...
    PaginatedChatSessionsResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        data: { type: array, items: { $ref: '#/components/schemas/ChatSession' } }
        meta:
          allOf:
            - $ref: '#/components/schemas/ResponseMeta'
            - type: object
              properties:
                pagination: { $ref: '#/components/schemas/PaginatedChatSessions' }
        error: { type: object, nullable: true, example: null }
    PaginatedChatMessagesResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        data: { type: array, items: { $ref: '#/components/schemas/ChatMessage' } }
        meta:
          allOf:
            - $ref: '#/components/schemas/ResponseMeta'
            - type: object
              properties:
                pagination: { $ref: '#/components/schemas/PaginatedChatMessages' }
        error: { type: object, nullable: true, example: null }
...